+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [vNext]
 ### Update
 - Cleanup cached data after check() execution. [Trello 1673](https://trello.com/c/CYbkzXia)
+- Using original view location on the screen for the Android scroll. [Trello 2138](https://trello.com/c/EI2mxmTD)
 ### Fixed
 - Didn't add escaping to css files when collecting dom for RCA. [Trello 2444](https://trello.com/c/yIez6VxK)